Add near real-time audio transcription feature with Whisper.cpp and Vosk models.
Whisper.ccp and Vosk are external librairies. See the README of whisper.cpp for hardware optimization.

Whisper.ccp is preferred for english and Vosk for french. Choose the model size according to your hardware (CPU, iGPU or GPU).

The transcription is made by a new MSFilter MSTranscript, that can use a Whipser.ccp model or a Vosk model. For Whisper.cpp, a real-time algorithm send the audio to the model as chunks of 3s duration, that overlap with the previous one. The transcibed words are sent to Liblinphone with successive events, and a LinphoneTranscription instance process them to get a sentence ready to be displayed. The application can display it thanks to the callback LinphoneTranscriptionDisplayCb. In conference mode, the name of the speaker is given with each sentence.

The transcription process in the mstranscript filter runs on a dedicated thread. The sampling rate of the audio must be 16000 Hz.

The ASR models can be easily added or updated thanks to the AbstractTranscript in the mstranscript filter.

TODO:
- manage the transcription duration in order to not increase the delay between spoken and transcibed words
- improve the correction of an already transcribed word
- set the procedure to download, build and install
- fix the build on macOS and Windows

fix missing file name